Move firebase call to top level
