+++ src/Views/Dashboard.js
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 import { Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 import * as actions from '../store/Actions/actionCreators';
-import firebase from "../config/firebaseConfig";
 
 import DashNav from '../Components/DashboardComponents/DashNav';
 import Profile from '../Components/DashboardComponents/Profile';
@@ -14,7 +13,6 @@ import UploadedArt from '../Components/UploadedArtComponent/UploadedArt'
 import Messaging from './Messaging';
 import MessageFulLView from '../Components/MessagingComponents/MessageFullView'
 
-const db = firebase.firestore();
 
 
 const DashboardContainer = styled.div`
@@ -46,31 +44,7 @@ const DashboardContainer = styled.div`
 		}
 	}
 `;
-function Dashboard (props){
-
-
-	useEffect(() => {
-		const fetchMessages = async () => {
-	
-		  try {
-			const snapshot = await db
-			  .collection("messages")
-			  .where('receiver_id', "==", props.loggedInUser._id)
-			  .get();
-	
-			const messages = snapshot.docs.map(x =>
-			  Object.assign({ id: x.id }, x.data(), {
-				sendAt: x.data().sendAt.toDate()
-			  })
-			);
-	
-			props.retrieveInboxMessages(messages);
-		  } catch (error) {
-			props.retrieveInboxMessages([]);
-		  }
-		};
-		fetchMessages();
-	  }, []);
+function Dashboard (){
 	return (
 		<DashboardContainer>
 			<div className="dashboard">
